+++ src/game/larasurf.c
@@ -292,7 +292,7 @@ void LaraSurfaceCollision(ITEM_INFO* item, COLL_INFO* coll)
         item->pos.y_rot -= 5 * PHD_DEGREE;
     }
 
-    int wh = GetWaterHeight(
+    int16_t wh = GetWaterHeight(
         item->pos.x, item->pos.y, item->pos.z, item->room_number);
     if (wh - item->pos.y <= -100) {
         item->goal_anim_state = AS_SWIM;
